+++ lib/src/inverted_index/async_index.dart
@@ -14,25 +14,35 @@ import 'package:text_indexing/text_indexing.dart';
 ///   index vocabulary also contains phrases up to [phraseLength] long,
 ///   concatenated from consecutive terms. The index size is increased by a
 ///   factor of [phraseLength];
-/// - [analyzer] is the [ITextAnalyzer] used to index the corpus terms;
+/// - [analyzer] is the [ITextAnalyzer] used to tokenize text for the index;
 /// - [vocabularyLength] is the number of unique terms in the corpus;
+/// - [zones] is a hashmap of zone names to their relative weight in the index;
+/// - [k] is the length of k-gram entries in the k-gram index;
 /// - [getDictionary] Asynchronously retrieves a [Dictionary] for a collection
 ///   of [Term]s from a [Dictionary] repository;
 /// - [upsertDictionary ] inserts entries into a [Dictionary] repository,
+///   overwriting any existing entries;
+/// - [getKGramIndex] Asynchronously retrieves a [KGramIndex] for a collection
+///   of [KGram]s from a [KGramIndex] repository;
+/// - [upsertKGramIndex ] inserts entries into a [KGramIndex] repository,
 ///   overwriting any existing entries;
 /// - [getPostings] asynchronously retrieves [Postings] for a collection
 ///   of [Term]s from a [Postings] repository;
 /// - [upsertPostings] inserts entries into a [Postings] repository,
 ///   overwriting any existing entries;
 /// - [getFtdPostings] return a [FtdPostings] for a collection of [Term]s from
-/// the [Postings], optionally filtered by minimum term frequency; and
+///   the [Postings], optionally filtered by minimum term frequency; and
 /// - [getIdFtIndex] returns a [IdFtIndex] for a collection of [Term]s from
-/// the [Dictionary].
+///   the [Dictionary];
 /// - [dictionaryLoader] asynchronously retrieves a [Dictionary] for a vocabulary
 ///   from a data source;
 /// - [dictionaryLengthLoader] asynchronously retrieves the number of terms in
 ///   the vocabulary (N);
 /// - [dictionaryUpdater] is callback that passes a [Dictionary] subset
+///    for persisting to a datastore;
+/// - [kGramIndexLoader] asynchronously retrieves a [KGramIndex] for a vocabulary
+///   from a data source;
+/// - [kGramIndexUpdater] is callback that passes a [KGramIndex] subset
 ///    for persisting to a datastore;
 /// - [postingsLoader] asynchronously retrieves a [Postings] for a vocabulary
 ///   from a data source; and
@@ -44,25 +54,26 @@ class AsyncCallbackIndex
   //
 
   @override
+  final int k;
+
+  @override
   final int phraseLength;
 
-  /// Asynchronously retrieves the number of terms in the vocabulary (N).
   @override
   final DictionaryLengthLoader dictionaryLengthLoader;
 
-  /// Asynchronously retrieves a [Dictionary] subset for a vocabulary from a
-  /// [Dictionary] data source, usually persisted storage.
   @override
   final DictionaryLoader dictionaryLoader;
 
-  /// A callback that passes a subset of a [Dictionary] containing new or
-  /// changed [DictionaryEntry] instances for persisting to the [Dictionary]
-  /// datastore.
   @override
   final DictionaryUpdater dictionaryUpdater;
 
-  /// Asynchronously retrieves a [Postings] subset for a vocabulary from a
-  /// [Postings] data source, usually persisted storage.
+  @override
+  final KGramIndexLoader kGramIndexLoader;
+
+  @override
+  final KGramIndexUpdater kGramIndexUpdater;
+
   @override
   final PostingsLoader postingsLoader;
 
@@ -72,13 +83,24 @@ class AsyncCallbackIndex
   final PostingsUpdater postingsUpdater;
 
   /// Instantiates a [AsyncCallbackIndex] instance:
+  /// - [phraseLength] is the maximum length of phrases in the index vocabulary.
+  ///   The minimum phrase length is 1. If phrase length is greater than 1, the
+  ///   index vocabulary also contains phrases up to [phraseLength] long,
+  ///   concatenated from consecutive terms. The index size is increased by a
+  ///   factor of [phraseLength];
   /// - [analyzer] is the [ITextAnalyzer] used to tokenize text for the index;
+  /// - [vocabularyLength] is the number of unique terms in the corpus;
   /// - [zones] is a hashmap of zone names to their relative weight in the index;
+  /// - [k] is the length of k-gram entries in the k-gram index;
   /// - [dictionaryLengthLoader] asynchronously retrieves the number of terms
   ///   in the vocabulary (N);
   /// - [dictionaryLoader] asynchronously retrieves a [Dictionary] for a vocabulary
   ///   from a data source;
   /// - [dictionaryUpdater] is callback that passes a [Dictionary] subset
+  ///    for persisting to a datastore;
+  /// - [kGramIndexLoader] asynchronously retrieves a [KGramIndex] for a vocabulary
+  ///   from a data source;
+  /// - [kGramIndexUpdater] is callback that passes a [KGramIndex] subset
   ///    for persisting to a datastore;
   /// - [postingsLoader] asynchronously retrieves a [Postings] for a vocabulary
   ///   from a data source; and
@@ -88,9 +110,12 @@ class AsyncCallbackIndex
       {required this.dictionaryLoader,
       required this.dictionaryUpdater,
       required this.dictionaryLengthLoader,
+      required this.kGramIndexLoader,
+      required this.kGramIndexUpdater,
       required this.postingsLoader,
       required this.postingsUpdater,
       required this.analyzer,
+      this.k = 3,
       this.zones = const <String, double>{},
       this.phraseLength = 1})
       : assert(phraseLength > 0, 'The phrase length must be 1 or greater');
@@ -113,6 +138,10 @@ class AsyncCallbackIndex
 ///   from a data source;
 /// - [dictionaryUpdater] is callback that passes a [Dictionary] subset
 ///    for persisting to a datastore;
+/// - [kGramIndexLoader] asynchronously retrieves a [KGramIndex] for a vocabulary
+///   from a data source;
+/// - [kGramIndexUpdater] is callback that passes a [KGramIndex] subset
+///    for persisting to a datastore;
 /// - [postingsLoader] asynchronously retrieves a [Postings] for a vocabulary
 ///   from a data source; and
 /// - [postingsUpdater] passes a [Postings] subset for persisting to a
@@ -123,6 +152,8 @@ class AsyncCallbackIndex
 /// - [vocabularyLength] calls [dictionaryLengthLoader].
 /// - [getDictionary] calls [dictionaryLoader];
 /// - [upsertDictionary ] calls [dictionaryUpdater];
+/// - [getKGramIndex] calls [kGramIndexLoader];
+/// - [upsertKGramIndex ] calls [kGramIndexUpdater];
 /// - [getPostings] calls [postingsLoader]; and
 /// - [upsertPostings] calls [postingsUpdater].
 abstract class AsyncCallbackIndexMixin implements InvertedIndex {
@@ -140,6 +171,14 @@ abstract class AsyncCallbackIndexMixin implements InvertedIndex {
   /// datastore.
   DictionaryUpdater get dictionaryUpdater;
 
+  /// Asynchronously retrieves a [KGramIndex] subset for a vocabulary from a
+  /// [KGramIndex] data source, usually persisted storage.
+  KGramIndexLoader get kGramIndexLoader;
+
+  /// A callback that passes a subset of a [KGramIndex] containing new or
+  /// changed entries for persisting to the [KGramIndex] repository.
+  KGramIndexUpdater get kGramIndexUpdater;
+
   /// Asynchronously retrieves a [Postings] subset for a vocabulary from a
   /// [Postings] data source, usually persisted storage.
   PostingsLoader get postingsLoader;
@@ -153,10 +192,17 @@ abstract class AsyncCallbackIndexMixin implements InvertedIndex {
       dictionaryLoader(terms);
 
   @override
-  Future<Postings> getPostings(Iterable<Term> terms) => postingsLoader(terms);
+  Future<void> upsertDictionary(Dictionary values) => dictionaryUpdater(values);
 
   @override
-  Future<void> upsertDictionary(Dictionary values) => dictionaryUpdater(values);
+  Future<KGramIndex> getKGramIndex([Iterable<Term>? terms]) =>
+      kGramIndexLoader(terms);
+
+  @override
+  Future<void> upsertKGramIndex(KGramIndex values) => kGramIndexUpdater(values);
+
+  @override
+  Future<Postings> getPostings(Iterable<Term> terms) => postingsLoader(terms);
 
   @override
   Future<void> upsertPostings(Postings values) => postingsUpdater(values);